Add files via upload
These are the deal class and and the actual game solitaire. The deal class deals the cards onto the table, and also puts the remainder deck after cards are dealed onto the table,
and Solitiaire.java is the actual creation of the game, and the back-end code of the game and gameplay.