Add files via upload
These are two classes. the Card class is an object creation class for the card. this sets the attributes of the card, namely the Suit and the rank of the card.
the Deck Class is an Array of the Card object in order to generate a full 52 pack array of unique cards. This also houses the shuffling method that shuffles the cards with a riffle shuffle and deals them into their respective piles.